Renaming to matcher and add type hints
