+++ python-threatexchange/threatexchange/extensions/video_vpdq/vpdq_brute_matcher.py
